started working on endgame ui

+++ ZombieGame/Game/Serializable/Score.cs
@@ -6,7 +6,7 @@ namespace ZombieGame.Game.Serializable
     public class Score
     {
         public string Name { get; set; }
-        public long Highscore { get; set; }
+        public float Highscore { get; set; }
         public DateTime Date { get; set; }
 
         public Score() { }
